rename default_hyperopt_loss file
