+++ freqtrade/optimize/hyperopt_loss_short_trade_dur.py
